requirements.txt
numpy
pandas
scikit-learn
nltk
matplotlib
seaborn
jupyter
streamlit
psycopg2-binary
beautifulsoup4
requests
lxml